+++ substrafl/logger.py
@@ -2,7 +2,14 @@ import logging
 
 
 def set_logging_level(loglevel):
-    """Set the logging level for substrafl
+    """Set the logging level for SubstraFL.
+    To change it, use the following code:
+
+    .. code-block:: python
+
+        import substrafl
+        import logging
+        substrafl.set_logging_level(loglevel=logging.DEBUG) # or any other log level
 
     Args:
         loglevel: log level (e.g. logging.INFO)